tests: don't error if path to remove doesn't exist

+++ pathops/tests/integration/juju/charms/machine/src/charm.py
@@ -50,6 +50,8 @@ class Charm(common.Charm):
 
     def remove_path(self, path: pathops.PathProtocol, recursive: bool = False) -> None:
         assert isinstance(path, pathops.LocalPath)
+        if not path.exists():
+            return
         if not path.is_dir():
             return path.unlink()
         if not recursive:
